webapp/requirements.txt
streamlit
pandas
matplotlib
Pillow
xarray
cfgrib
eccodes-python
scikit-learn
xgboost
numpy